Internal #861: Refactor WindowDistinctAggregator
Move the WindowDistinctAggregator class to separate files.
