+++ src/function/window/window_aggregator.cpp
@@ -1,4 +1,5 @@
 #include "duckdb/function/window/window_aggregator.hpp"
+
 #include "duckdb/function/window/window_collection.hpp"
 #include "duckdb/function/window/window_shared_expressions.hpp"
 #include "duckdb/planner/expression/bound_window_expression.hpp"
